[JL] refactor db and s3 seed script into multiple files
