+++ prisma/helpers/createUser.ts
@@ -1,0 +1,11 @@
+import { users } from "@prisma/client"
+import { faker } from "./faker"
+
+export const createUser = (): users => {
+  return {
+    uid: faker.string.uuid(),
+    name: faker.person.fullName(),
+    email: faker.internet.email(),
+    contact_details: {},
+  }
+}
