+++ prisma/helpers/faker.ts
@@ -1,0 +1,4 @@
+import { faker } from "@faker-js/faker"
+faker.seed(12451)
+
+export { faker }
